New Server #11 #5
A server capable of connecting 2 players and starting a game without them having to exchange information like IP address and ports to use. Also fixing a bug that caused problems when restarting the game with the other player's decisions, in addition to adding wait times to make game entry more realistic.